+++ kmp-maps-compose/src/wasmJsMain/kotlin/eu/buney/maps/MapResize.wasmJs.kt
@@ -1,0 +1,12 @@
+@file:OptIn(ExperimentalWasmJsInterop::class)
+
+package eu.buney.maps
+
+import kotlin.js.ExperimentalWasmJsInterop
+
+internal actual fun triggerMapResize(map: NativeMap) {
+    triggerResize(map.handle)
+}
+
+private fun triggerResize(map: eu.buney.maps.jsinterop.GMap): Unit =
+    js("google.maps.event.trigger(map, 'resize')")